impr: Various fixes and improvements to the pattern editor (#2559)
- fixed crash when utf8 chars were present in text editor
- fixed unable to scroll when cursor at line 1
- removed dependencies on thext editor that were not being used.

I had to go back to the old code (old for me) and fit in the changes
that were applied to the new code.That was only possible by
incorporating some of the new structural differences to the text editor.
This created new bugs and crashes that I ve have fixed but there may be
ones that I couldn't find in the very small amount of time I could spend
testing so that this commit wouldn't be delayed. If more crashes are
found due to the mixing of old and new code they should be resolved when
the new code is brought in.

(cherry picked from commit bfa97880999a9956156ba6d1cba6dcfd493180cd)

+++ plugins/builtin/source/content/text_highlighting/pattern_language.cpp
@@ -1247,7 +1247,7 @@ namespace hex::plugin::builtin {
                 if (isLocationValid(error.getLocation())) {
                     auto key = ui::TextEditor::Coordinates(error.getLocation().line, error.getLocation().column);
 
-                    if (!errorMarkers.contains(key) || errorMarkers[key].first < error.getLocation().length)
+                    if (!errorMarkers.contains(key) || errorMarkers[key].first < (i32) error.getLocation().length)
                         errorMarkers[key] = std::make_pair(error.getLocation().length, processMessage(error.getMessage()));
                 }
             }
